Correções, ajuste e validações nas entidades,DTOs e services

+++ src/main/java/br/com/sicredi/votacao/dto/VotoDTO.java
@@ -1,7 +1,5 @@
 package br.com.sicredi.votacao.dto;
 
-import java.math.BigInteger;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class VotoDTO {
 	
-    private BigInteger idPauta;
+    private Long idPauta;
     private Long numeroCpf;
     private Integer indicadorVotoSim;
