+++ src/main/java/br/com/sicredi/votacao/repository/PautaRepository.java
@@ -1,12 +1,10 @@
 package br.com.sicredi.votacao.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.sicredi.votacao.model.Pauta;
 
-public interface PautaRepository extends JpaRepository<Pauta, BigInteger> {
+public interface PautaRepository extends JpaRepository<Pauta, Long> {
 
     Pauta findByDescricaoPautaIgnoreCase(String desPauta);
